+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
 - master
-  - support setting custom faces in object mode
+  - A region can be selected by repeating the same composable command
+    twice.
+  - Support setting custom faces in object mode
 - 0.0.1
-  - initial stable release
+  - Initial stable release